Do not install nvm if already installed

+++ setup/install_node.sh
@@ -6,8 +6,10 @@ source ~/dotfiles/terminal/bash/exports.sh
 
 echo "Installing node..."
 
-curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.38.0/install.sh | bash
-mkdir -p ~/.nvm
+if [ ! -e ~/.nvm ]; then
+	curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.38.0/install.sh | bash
+	mkdir -p ~/.nvm
+fi
 source ~/dotfiles/terminal/bash/load_nvm.sh
 nvm install 6.15.1
 nvm install 8.9.4
